+++ packages/mapboxgl-map-adapter/src/layer-adapters/BaseAdapter.ts
@@ -1,11 +1,12 @@
-import { Map } from 'mapbox-gl';
+import { Map } from 'maplibre-gl';
 import { MainLayerAdapter, AdapterOptions } from '@nextgis/webmap';
 import { TLayer } from '../MapboxglMapAdapter';
 
 let ID = 0;
 
 export abstract class BaseAdapter<O extends AdapterOptions = AdapterOptions>
-  implements MainLayerAdapter<Map, TLayer, O> {
+  implements MainLayerAdapter<Map, TLayer, O>
+{
   layer?: TLayer;
   map?: Map;
   protected readonly _layerId: string;