+++ packages/mapboxgl-map-adapter/src/layer-adapters/ImageAdapter.ts
@@ -1,12 +1,15 @@
 import { MainLayerAdapter, ImageAdapterOptions } from '@nextgis/webmap';
-import { Map } from 'mapbox-gl';
+import { Map } from 'maplibre-gl';
 import { TLayer } from '../MapboxglMapAdapter';
 import { BaseAdapter } from './BaseAdapter';
 
 export class ImageAdapter
   extends BaseAdapter<ImageAdapterOptions>
-  implements MainLayerAdapter<Map, TLayer, ImageAdapterOptions> {
-  addLayer(options: ImageAdapterOptions): string[] | undefined {
+  implements MainLayerAdapter<Map, TLayer, ImageAdapterOptions>
+{
+  addLayer(
+    options: ImageAdapterOptions & { before?: string },
+  ): string[] | undefined {
     if (this.options) {
       options = { ...this.options, ...options };
     }
@@ -14,9 +17,10 @@ export class ImageAdapter
     const url = options.url;
     if (url && this.map) {
       if (options.subdomains) {
-        tiles = (typeof options.subdomains === 'string'
-          ? options.subdomains.split('')
-          : options.subdomains
+        tiles = (
+          typeof options.subdomains === 'string'
+            ? options.subdomains.split('')
+            : options.subdomains
         ).map((x) => {
           const subUrl = url.replace('{s}', x);
           return subUrl;
@@ -47,7 +51,6 @@ export class ImageAdapter
           },
           paint: {},
         },
-        // @ts-ignore
         options.before,
       );
       this.layer = [this._layerId];