+++ packages/webmap/src/interfaces/RuntimeParams.ts
@@ -1,6 +1,4 @@
-/**
- * @public
- */
+
 export interface RuntimeParams {
   params(): { [paramName: string]: any };
   get(name: string): any;